+++ lib/capistrano/redmine/deployment/gem_version.rb
@@ -10,8 +10,8 @@ module Capistrano
 
       module VERSION
         MAJOR = 1
-        MINOR = 0
-        TINY  = 1
+        MINOR = 1
+        TINY  = 0
         PRE   = nil
 
         STRING = [MAJOR, MINOR, TINY, PRE].compact.join(".")